+++ client/components/index.js
@@ -17,3 +17,5 @@ export { default as ChangeTheme } from './ChangeTheme';
 export { default as OneArticle } from  './OneArticle';
 export { default as ReaderNav } from './ReaderNav';
 export { default as MyProfile } from './MyProfile';
+export { default as FullCard } from './FullCard'
+export { default as GeneralCardIcons } from './GeneralCardIcons'
