+++ client/components/main.js
@@ -4,7 +4,7 @@ import { connect } from 'react-redux'
 import { withRouter, Link } from 'react-router-dom'
 import { logout } from '../store'
 import { AddPopup } from './index'
-import { Button, Icon, Divider, Image } from 'semantic-ui-react'
+import { Button, Icon, Divider } from 'semantic-ui-react'
 import { ChangeTheme } from '../components'
 /**
  * COMPONENT
