+++ package.json
@@ -38,6 +38,7 @@
     "pg-hstore": "^2.3.2",
     "prop-types": "^15.5.8",
     "react": "^16.0.0",
+    "react-dnd-html5-backend": "^2.5.4",
     "react-dom": "^16.0.0",
     "react-html-parser": "^2.0.1",
     "react-mosaic-component": "^1.0.3",
